Skip rxing messages that self node transmits
Note that this will result in silently ignored RX messages when those come
from the target node itself. This is fine when using the "*" rx directive in
compose, but would be pretty confusing if you explicitly named a message and
didn't see anything at all show up in RX codegen for it. We should emit a warning.